+++ crankshaft-engine/src/service/runner/backend.rs
@@ -5,9 +5,10 @@ use std::process::ExitStatus;
 
 use anyhow::Result;
 use async_trait::async_trait;
+use crankshaft_monitor::proto::Event;
 use futures::future::BoxFuture;
 use nonempty::NonEmpty;
-use tokio::sync::oneshot;
+use tokio::sync::broadcast;
 use tokio_util::sync::CancellationToken;
 
 use crate::Task;
@@ -41,15 +42,15 @@ pub trait Backend: Debug + Send + Sync + 'static {
 
     /// Runs a task in a backend.
     ///
-    /// The optional `started` channel is notified when the first execution of
-    /// the task has started.
+    /// The optional event_sender sends the task lifecycle events to any client
+    /// connected to `Crankshaft`
     ///
     /// Returns a collection of exit status corresponding to the task's
     /// executions.
     fn run(
         &self,
         task: Task,
-        started: Option<oneshot::Sender<()>>,
+        event_sender: Option<broadcast::Sender<Event>>,
         token: CancellationToken,
     ) -> Result<BoxFuture<'static, Result<NonEmpty<ExitStatus>, TaskRunError>>>;
 }