+++ crankshaft-monitor/src/lib.rs
@@ -1,5 +1,7 @@
 //! Crate for monitoring crankshaft events.
 use std::net::SocketAddr;
+use std::time::SystemTime;
+use std::time::UNIX_EPOCH;
 
 use anyhow::Result;
 use proto::Event;
@@ -29,6 +31,52 @@ pub fn start_monitoring(addr: SocketAddr) -> Result<(broadcast::Sender<Event>, J
             .serve(addr)
             .await
     });
-
     Ok((event_sender, server_handle))
+}
+
+/// current timestamp as i64
+pub fn now_millis() -> i64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("System time before UNIX epoch")
+        .as_millis() as i64
+}
+
+/// Sends an event through a broadcast channel.
+///
+/// No event is sent if the specified broadcast channel is `None`.
+#[macro_export]
+macro_rules! send_event {
+    ($sender:expr, $task_id:expr, $event_type:expr, $message:literal) => {
+        if let Some(sender) = $sender.as_ref() {
+            let _ = sender.send($crate::proto::Event {
+                task_id: $task_id.to_owned(),
+                event_type: $event_type as i32,
+                timestamp: $crate::now_millis(),
+                message: $message.to_string(),
+            });
+        }
+    };
+    ($sender:expr, $task_id:expr, $event_type:expr, $fmt:literal, $($arg:tt)*) => {
+        if let Some(sender) = $sender.as_ref() {
+            let message = format!($fmt, $($arg)*);
+            let _ = sender.send($crate::proto::Event {
+                task_id: $task_id.to_owned(),
+                event_type: $event_type as i32,
+                timestamp: $crate::now_millis(),
+                message,
+            });
+        }
+    };
+    ($sender:expr, $task_id:expr, $event_type:expr, $($arg:tt)*) => {
+        if let Some(sender) = $sender.as_ref() {
+            let message = format!("{}",$($arg)*);
+            let _ = sender.send($crate::proto::Event {
+                task_id: $task_id.to_owned(),
+                event_type: $event_type as i32,
+                timestamp: $crate::now_millis(),
+                message,
+            });
+        }
+    };
 }